dcp-815 temporarily rename metadata to scratch
