+++ tests/manifest/test_generator.py
@@ -4,7 +4,7 @@ from hca_ingest.api.ingestapi import IngestApi
 from mock import MagicMock
 
 from exporter.graph.graph_crawler import GraphCrawler
-from exporter.metadata import MetadataService
+from exporter.scratch import MetadataService
 from manifest.generator import ManifestGenerator
 from tests.mocks.files import MockEntityFiles
 from tests.mocks.ingest import MockIngestAPI
